adding luminosity tests and exporting all functions
