+++ packages/fast-colors/src/index.ts
@@ -1,1 +1,3 @@
+export { contrast } from "./contrast";
+export { luminance } from "./luminosity";
 export * from "./range";
